+++ src/redux/reducers/settings.reducer.js
@@ -1,17 +1,6 @@
 import { combineReducers } from 'redux';
 
 
-const allSettingsReducer = (state = {}, action) => {
-  switch (action.type) {
-    case 'SET_ALL_SETTINGS':
-      return action.payload;
-    case 'UNSET_ALL_SETTINGS':
-      return {};
-    default:
-      return state;
-  }
-};
-
 const scrollingReducer = (state = {canScroll: false}, action) => {
   switch (action.type) {
     case 'SET_SCROLL':
@@ -23,6 +12,5 @@ const scrollingReducer = (state = {canScroll: false}, action) => {
 };
 
 export default combineReducers({
-  allSettingsReducer,
   scrollingReducer,
 });
