+++ src/redux/sagas/account.saga.js
@@ -1,20 +1,6 @@
 import { put, takeLatest } from 'redux-saga/effects';
 import axios from 'axios';
 
-
-// // toggle active status of product
-// function* toggleActiveProduct(action) {
-//   try {
-//     yield axios.put(`/api/account/products`, action.payload);
-//     yield put({ type: 'FETCH_PRODUCTS' });
-//   } catch (error) {
-//     // console.log('PUT products route has failed', error);
-//     if (error?.response?.status === 403) {
-//       yield put({ type: 'UNSET_USER' });
-//     }
-//   }
-// }
-
 function* fetchErrors() {
   try {
     const response = yield axios.get(`/api/account/errors`);
