+++ src/redux/sagas/settings.saga.js
@@ -3,66 +3,6 @@ import axios from 'axios';
 
 
 
-function* getAllSettings() {
-  try {
-    const response = yield axios.get(`/api/settings`);
-    yield put({ type: 'SET_ALL_SETTINGS', payload: response.data })
-  } catch (error) {
-    console.log('getting all settings has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
-
-function* loopSpeed(action) {
-  try {
-    yield axios.put(`/api/settings/loopSpeed`, action.payload);
-    yield put({ type: 'FETCH_SETTINGS' })
-  } catch (error) {
-    console.log('setting bot speed has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
-
-function* fullSync(action) {
-  try {
-    yield axios.put(`/api/settings/fullSync`, action.payload);
-    yield put({ type: 'FETCH_SETTINGS' })
-  } catch (error) {
-    console.log('setting bot speed has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
-
-function* syncQuantity(action) {
-  try {
-    yield axios.put(`/api/settings/orderSyncQuantity`, action.payload);
-    yield put({ type: 'FETCH_SETTINGS' })
-  } catch (error) {
-    console.log('setting bot speed has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
-
-function* toggleMaintenance(action) {
-  try {
-    yield axios.put(`/api/settings/toggleMaintenance`, action.payload);
-    yield put({ type: 'FETCH_SETTINGS' })
-  } catch (error) {
-    console.log('setting bot speed has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
-
 function* bulkPairRatio(action) {
   try {
     yield axios.put(`/api/settings/bulkPairRatio`, action.payload);
@@ -136,29 +76,6 @@ function* killLock(action) {
   }
 }
 
-function* ordersReset() {
-  try {
-    yield axios.post(`/api/settings/ordersReset`);
-    // yield put({ type: 'FETCH_ORDERS' });
-  } catch (error) {
-    console.log('post account route factoryReset has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
-
-function* factoryReset() {
-  try {
-    yield axios.post(`/api/settings/factoryReset`);
-    yield put({ type: 'UNSET_USER' });
-  } catch (error) {
-    console.log('post account route factoryReset has failed', error);
-    if (error.response.status === 403) {
-      yield put({ type: 'UNSET_USER' });
-    }
-  }
-}
 
 function* test() {
   try {
@@ -173,19 +90,12 @@ function* test() {
 }
 
 function* settingsSaga() {
-  yield takeLatest('FETCH_SETTINGS', getAllSettings);
-  yield takeLatest('SEND_LOOP_SPEED', loopSpeed);
-  yield takeLatest('SEND_FULL_SYNC', fullSync);
-  yield takeLatest('SEND_SYNC_QUANTITY', syncQuantity);
   yield takeLatest('SET_BULK_PAIR_RATIO', bulkPairRatio);
   yield takeLatest('SET_PROFIT_ACCURACY', sendProfitAccuracy);
   yield takeLatest('SET_MAX_TRADE_LOAD', sendTradeLoadMax);
   yield takeLatest('POST_MAX_REINVEST_RATIO', postMaxReinvestRatio);
   yield takeLatest('SAVE_RESERVE', reserve);
-  yield takeLatest('TOGGLE_MAINTENANCE', toggleMaintenance);
   yield takeLatest('KILL_LOCK', killLock);
-  yield takeLatest('ORDERS_RESET', ordersReset);
-  yield takeLatest('FACTORY_RESET', factoryReset);
   yield takeLatest('TEST', test);
 }
